refactor: moved sqsdomain orderbook references to domain
